move doesFolderExists to common api
