+++ src/api/common.ts
@@ -1,4 +1,5 @@
 import { ipcRenderer } from 'electron';
+import fs from 'promise-fs';
 
 export const sendNotification = (content: string) => {
     ipcRenderer.send('send-notification', content);
@@ -9,3 +10,12 @@ export const showOnTray = (content: string) => {
 export const reloadWindow = () => {
     ipcRenderer.send('reload-window');
 };
+
+export async function doesFolderExists(dirPath: string) {
+    return await fs
+        .stat(dirPath)
+        .then((stats) => {
+            return stats.isDirectory();
+        })
+        .catch(() => false);
+}
